Build boost::iostreams with the zlib filter
Explicitly add zlib to the requirements so that
boost::iostreams gets built with the zlib filter.
This was being pulled in implicitly at one point
but recently stopped working; make it explicit.
Add a compile+link test to make sure the final
library includes the necessary symbols. Closes #122.

This is essentially the same as PR #115 (which
fixed the issue for the 1.74.x branch) but for
the main branch.